requirements.txt
pypdf==3.15.0
pillow==10.0.0
reportlab==4.0.4